requirements.txt
openai
python-dotenv
maturin
fastapi
uvicorn